Factored UserResource and profile into organisation/api.py.
